% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{addRemoveEdges}
\alias{addRemoveEdges}
\alias{addRemove}
\alias{add}
\title{addRemoveEdges(graphAddRemove)}
\usage{
addRemoveEdges(graphAddRemove)
addRemoveEdges(dyncomm,graphAddRemove)
addRemove(dyncomm,graphAddRemove)
add(dyncomm,graphAddRemove)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the 
function call}

\item{graphAddRemove}{Either the matrix or the filename that contains the 
edges to add/remove}
}
\value{
FALSE if any kind of error occurred. Otherwise, TRUE
}
\description{
This method reads edges from either a matrix or a file and adds or removes 
them from the graph.
}
\details{
The matrix must have at least two columns with the source and destination
vertices. Optionally, it can have a third column with the weight of the edge.

The file must have only one edge per line, with values separated by a white
space (both SPACE and TAB work in any amount and combination).

The first value is the source vertex, the second is the destination vertex, 
and the third is the weight.

The weight can be ommited if the edge is to be added using the default weight
of 1 (one), or if the parameter to ignore weights was set.

If the weight is exactly zero, the edge is removed from the graph.

If a vertex, mentioned in the source or destination, does not exist it will be 
added to the graph.

The method detects automatically if the weight is present on a row by row 
basis so some rows may have weights defined and others not.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$addRemoveEdges("graphAddRemoveFile.txt")
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
addRemoveEdges(dc,"graphAddRemoveFile.txt")
}

}
\author{
poltergeist0
}
