% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{DynComm$addRemoveEdgesFile}
\alias{DynComm$addRemoveEdgesFile}
\alias{addRemoveEdgesFile}
\title{addRemoveEdgesFile(graphAddRemoveFile)}
\usage{
addRemoveEdgesFile(graphAddRemoveFile)
}
\value{
FALSE if any kind of error occurred. Otherwise, TRUE
}
\description{
This method reads edges from a file and adds or removes them from the graph.

The file must have only one edge per line, with values separated by a white
space (both SPACE and TAB work in any amount and combination).

The first value is the source node, the second is the destination node, and 
the third is the weight.

The weight can be ommited if the edge is to be added using the default weight
of 1 (one), or if the parameter to ignore weights was set.

If the weight is exactly zero, the edge is removed from the graph.

If a node, mentioned in the source or destination, does not exist it will be 
added to the graph.

The method detects automatically if the weight is present on a row by row 
basis so some rows may have weights defined and others not.
}
\examples{
dc<-DynComm(ALGORITHM$LOUVAIN,QUALITY$MODULARITY,parameters)
dc$addRemoveEdgesFile("graphAddRemoveFile.txt")

}
