% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{communitiesEdgeCount}
\alias{communitiesEdgeCount}
\alias{communitiesEdgeCount.DynComm}
\title{communitiesEdgeCount}
\usage{
communitiesEdgeCount(x, ...)
}
\arguments{
\item{x}{A DynComm object, if not using the inline version of the
function call}

\item{...}{Any future additional parameters}
}
\value{
the number of community to community edges in the graph
}
\description{
This method returns the number of community to community edges in the graph
from the selected post processing algorithm or the main algorithm, after the
last iteration.
}
\examples{
\donttest{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$communitiesEdgeCount()
}
\donttest{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
communitiesEdgeCount(dc)
}

}
\seealso{
\code{\link{DynComm}} , \code{\link{postProcess}}
}
\author{
poltergeist0
}
