% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{DynComm$community}
\alias{DynComm$community}
\alias{community}
\title{community(node)}
\usage{
community(node)
}
\arguments{
\item{node}{The name of the intended node}
}
\value{
an unsigned integer with the community of the given node
}
\description{
Get the community of the given node after the last iteration.
}
\examples{
dc<-DynComm(ALGORITHM$LOUVAIN,QUALITY$MODULARITY,parameters)
dc$community(8)

}
