% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{DynComm$communityInnerEdgesWeight}
\alias{DynComm$communityInnerEdgesWeight}
\alias{communityInnerEdgesWeight}
\title{communityInnerEdgesWeight(community)}
\usage{
communityInnerEdgesWeight(community)
}
\arguments{
\item{community}{The name of the intended community}
}
\value{
a floating point number with the weight
}
\description{
Get the sum of weights of the inner edges of the given community after the 
last iteration.
}
\examples{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$communityInnerEdgesWeight(1)

}
