% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{DynComm$communityMapping}
\alias{DynComm$communityMapping}
\alias{communityMapping}
\title{communityMapping()}
\usage{
communityMapping()
}
\value{
a two column matrix with communities in the first column and the nodes in the second
}
\description{
Get the community mapping for all communities after the last iteration.
}
\examples{
dc<-DynComm(ALGORITHM$LOUVAIN,QUALITY$MODULARITY,parameters)
dc$communityMapping()

}
