% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{communityNeighbours}
\alias{communityNeighbours}
\alias{DynComm.communityNeighbours}
\title{communityNeighbours(community)}
\usage{
DynComm.communityNeighbours(dyncomm,community)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the 
function call}

\item{community}{The community to get neighbours from}
}
\value{
a matrix of all communities in the graph that are neighbours of the 
given community and their edge weight
}
\description{
Get all neighbours (communities connected through direct edges) of the given 
community in the graph from the selected post processing algorithm or the main 
algorithm, after the last iteration.
}
\details{
The return value is a matrix with two columns. The first is the neighbour and
the second is the weight of the edge that connects them.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$communityNeighbours(community)
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
communityNeighbours(dc,community)
}

}
\seealso{
\code{\link{DynComm}} , \code{\link{postProcess}}
}
\author{
poltergeist0
}
