% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{DynComm$communityNodeCount}
\alias{DynComm$communityNodeCount}
\alias{communityNodeCount}
\title{communityNodeCount(community)}
\usage{
communityNodeCount(community)
}
\arguments{
\item{community}{The name of the intended community}
}
\value{
an unsigned integer with the number nodes in the given community
}
\description{
Get the amount of nodes in the given community after the last iteration.
}
\examples{
dc<-DynComm(ALGORITHM$LOUVAIN,QUALITY$MODULARITY,parameters)
dc$communityNodeCount(3)

}
