% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{DynComm$communityTotalWeight}
\alias{DynComm$communityTotalWeight}
\alias{communityTotalWeight}
\title{communityTotalWeight(community)}
\usage{
communityTotalWeight(community)
}
\arguments{
\item{community}{The name of the intended community}
}
\value{
a floating point number with the weight
}
\description{
Get the sum of weights of all edges of the given community after the last iteration.
}
\examples{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$communityTotalWeight(1)

}
