% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{communityVertexCount}
\alias{communityVertexCount}
\alias{DynComm.communityVertexCount}
\alias{DynComm.communityNodeCount}
\title{communityVertexCount(community)}
\usage{
DynComm.communityVertexCount(dyncomm,community)
DynComm.communityNodeCount(dyncomm,community)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the 
function call}

\item{community}{The name of the intended community}
}
\value{
an unsigned integer with the number of vertices in the given community
}
\description{
Get the amount of vertices in the given community from the selected post 
processing algorithm or the main algorithm, after the last iteration.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$communityVertexCount(3)
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
communityVertexCount(dc,3)
}

}
\seealso{
\code{\link{DynComm}} , \code{\link{postProcess}}
}
\author{
poltergeist0
}
