% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{data}
\name{CRITERION}
\alias{CRITERION}
\alias{criterion}
\alias{Criterion}
\title{List of available CRITERION (quality measurement functions).}
\format{A named list with the names of the available CRITERION:
\describe{
  \item{MODULARITY}{Newman-Girvan}
}}
\usage{
CRITERION
}
\description{
A criterion is used to indicate the proximity of the current grouping 
of vertices (communities) to the optimum one. Theoretically, the bigger 
the value, the closer the current grouping is to the best possible 
grouping.

Each CRITERION internally defines two functions. One is used to 
evaluate if moving a vertex from one group (community) to another 
possibly yields a better overall result. The other is used to measure 
the actual overall quality of the entire grouping (current community 
mapping).
}
\examples{
CRITERION$MODULARITY

}
\keyword{datasets}
