% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityOPT2.R
\name{densopt}
\alias{densopt}
\title{The implementation of eTILES algorithm.}
\usage{
densopt(graph, graph.directed = TRUE, comms = NULL,
  type.names = c("num", "alfa"))
}
\arguments{
\item{graph}{input edge list graph}

\item{graph.directed}{is graph directed or not}

\item{comms}{community node assignment for each node in graph}

\item{type.names}{are node names numeric or alfanumeric}
}
\value{
dataframe with nodes and new assigned community
}
\description{
The implementation of eTILES algorithm.
}
\examples{
str_length(letters)
}
\seealso{
\code{\link{nchar}} which this function wraps
}
