% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{package}
\name{DynComm-package}
\alias{DynComm-package}
\alias{dyncomm-package}
\alias{Dyncomm-package}
\title{DynComm: Dynamic Network Communities Detection}
\description{
Bundle of algorithms used for evolving network analysis regarding 
community detection.
Implements several algorithms, using a common API, that calculate 
communities for graphs whose vertices and edges change over time.
Edges, which can have new vertices, can be added or deleted, and changes in the 
communities are calculated without recalculating communities for the 
entire graph.
}
