% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{edgeCount}
\alias{edgeCount}
\alias{DynComm.edgeCount}
\title{edgeCount()}
\usage{
DynComm.edgeCount(dyncomm)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the 
function call}
}
\value{
the number of vertex to vertex edges in the graph
}
\description{
This method returns the number of vertex to vertex edges in the graph from the
selected post processing algorithm or the main algorithm, after the last 
iteration.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$edgeCount()
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
edgeCount(dc)
}

}
\seealso{
\code{\link{DynComm}} , \code{\link{postProcess}}
}
\author{
poltergeist0
}
