% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{edgeWeight}
\alias{edgeWeight}
\alias{edgeWeight.DynComm}
\title{edgeWeight}
\usage{
edgeWeight(x,source,destination, ...)
}
\arguments{
\item{x}{A DynComm object, if not using the inline version of the
function call}

\item{source}{The name of the source vertex that is part of the edge}

\item{destination}{The name of the destination vertex that is part of the edge}

\item{...}{Any future additional parameters}
}
\value{
a floating point number with the weight
}
\description{
Get the weight of the edge that goes from source vertex to destination vertex
from the selected post processing algorithm or the main algorithm, after the
last iteration.
}
\examples{
\donttest{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$edgeWeight(12,42)
}
\donttest{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
edgeWeight(dc,12,42)
}

}
\seealso{
\code{\link{DynComm}} , \code{\link{postProcess}}
}
\author{
poltergeist0
}
