% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eTILES.R
\name{etiles}
\alias{etiles}
\title{The implementation of eTILES algorithm.}
\usage{
etiles(streamfile.edge.removal, init.graph = NULL, obs = 7,
  path = "", start = NULL, end = NULL)
}
\arguments{
\item{streamfile.edge.removal}{.csv file with stream input edges to remove from initial graph}

\item{init.graph}{input initial edge list graph}

\item{obs}{observation window (days)}

\item{path}{Path where generate the results and find the edge file}

\item{start}{starting date}

\item{end}{ending date}
}
\value{
dynamic community detection represented in a XXXXXX type of object
}
\description{
The implementation of eTILES algorithm.
}
\examples{

}
\seealso{
\code{\link{nchar}} which this function wraps
}
