% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{DynComm$getAlgorithm}
\alias{DynComm$getAlgorithm}
\alias{getAlgorithm}
\title{getAlgorithm()}
\usage{
getAlgorithm()
}
\value{
ALGORITHM
}
\description{
This method returns the algorithm with which this object was initialized.
}
\examples{
dc<-DynComm(ALGORITHM$LOUVAIN,QUALITY$MODULARITY,parameters)
dc$getAlgorithm()

}
