% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{data}
\name{QUALITY}
\alias{QUALITY}
\title{QUALITY}
\format{A named list with the names of the available algorithms:
 \describe{
  \item{quality}{See available quality below.}
 }}
\usage{
QUALITY$quality
}
\description{
List of available quality measurement functions.
}
\details{
A quality measurement function mainly returns a value that indicates 
the proximity of the current grouping of nodes (communities) to the
optimum one. Theoretically, the bigger the value, the closer the 
current grouping is to the best possible grouping.

Each QUALITY internally defines two functions. One is used to 
evaluate if moving a node from one group (community) to another 
possibly yields a better overall result. The other is used to measure 
the actual overall quality of the entire grouping (current community 
mapping).
}
\section{Currently supported quality measurement functions are}{

\describe{
  \item{MODULARITY}{TO DO}
  \item{BALMOD}{TO DO}
}
}

\section{Quality measurement functions expected to be supported in the future}{

}

\examples{
QUALITY$MODULARITY
QUALITY$BALMOD

}
\keyword{datasets}
