% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDyn.R
\name{rdyn}
\alias{rdyn}
\title{The implementatio of TILES algorithm.}
\usage{
rdyn(size = 1000, iterations = 100, avg_deg = 15, sigma = 0.6,
  lambdad = 1, alpha = 2.5, paction = 1, prenewal = 0.8,
  quality_threshold = 0.2, new_node = 0, del_node = 0,
  max_evts = 1)
}
\arguments{
\item{size}{input initial size of graph}

\item{iterations}{number of iterations}

\item{avg_deg}{node average degree}

\item{sigma}{XXXX}

\item{lambdad}{XXXX}

\item{alpha}{XXXXX}

\item{paction}{XXXXX}

\item{prenewal}{XXXXX}

\item{quality_threshold}{XXXX}

\item{new_node}{XXXXX}

\item{del_node}{XXXXX}

\item{max_evts}{XXXXX}
}
\value{
dynamically generated network in a stream of edges
}
\description{
The implementatio of TILES algorithm.
}
\examples{
str_length(letters)
}
\seealso{
\code{\link{nchar}} which this function wraps
}
