% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TILES.R
\name{tiles}
\alias{tiles}
\title{The implementation of TILES algorithm.}
\usage{
tiles(streamfile, init.graph = NULL, ttl = Inf, obs = 7, path = "",
  start = NULL, end = NULL)
}
\arguments{
\item{streamfile}{.csv file with stream input edges}

\item{init.graph}{input initial edge list graph}

\item{ttl}{edge time to live (days)}

\item{obs}{observation window (days)}

\item{path}{Path where generate the results and find the edge file}

\item{start}{starting date}

\item{end}{ending date}
}
\value{
dynamic community detection represented in a XXXXXX type of object
}
\description{
The implementation of TILES algorithm.
}
\examples{
str_length(letters)
}
\seealso{
\code{\link{nchar}} which this function wraps
}
