% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{DynComm$time}
\alias{DynComm$time}
\alias{time}
\title{time()}
\usage{
time()
time(DynComm)
}
\value{
an unsigned integer with the total processing time
}
\description{
Get the cumulative time spent on processing after the last iteration.
}
\examples{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$time()
## 2.3
}
