% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{DynComm$vertices}
\alias{DynComm$vertices}
\alias{vertices}
\title{vertices(community)}
\usage{
vertices(community)
}
\arguments{
\item{community}{The name of the intended community}
}
\value{
a list of vertices belonging to the given community
}
\description{
Get all vertices belonging to the given community after the last iteration.
}
\examples{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$vertices(6)

}
