% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{verticesAll}
\alias{verticesAll}
\alias{nodesAll}
\title{verticesAll()}
\usage{
verticesAll()
verticesAll(dyncomm)
nodesAll()
nodesAll(dyncomm)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the function call}
}
\value{
a list of all vertices in the graph
}
\description{
Get all vertices in the graph after the last iteration.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$verticesAll()
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
verticesAll(dc)
}

}
\author{
poltergeist0
}
