% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{communities}
\alias{communities}
\alias{communities.DynComm}
\title{communities}
\usage{
communities(x, ...)
}
\arguments{
\item{x}{A DynComm object, if not using the inline version of the
function call}

\item{...}{Any future additional parameters}
}
\value{
a list of all communities
}
\description{
This method returns all communities from the selected post processing
algorithm or the main algorithm, after the last iteration.
}
\examples{
\donttest{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$communities()
}
\donttest{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
communities(dc)
}

}
\seealso{
\code{\link{DynComm}} , \code{\link{postProcess}}
}
\author{
poltergeist0
}
