% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{DynComm$communities}
\alias{DynComm$communities}
\alias{communities}
\title{communities()}
\usage{
communities()
}
\value{
a list of all communities
}
\description{
This method returns all communities after the last iteration.
}
\examples{
dc<-DynComm(ALGORITHM$LOUVAIN,QUALITY$MODULARITY,parameters)
dc$communities()

}
