% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{communities}
\alias{communities}
\title{communities()}
\usage{
communities()
communities(dyncomm)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the function call}
}
\value{
a list of all communities
}
\description{
This method returns all communities after the last iteration.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$communities()
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
communities(dc)
}

}
\author{
poltergeist0
}
