% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{communitiesEdgeCount}
\alias{communitiesEdgeCount}
\alias{DynComm.communitiesEdgeCount}
\title{communitiesEdgeCount()}
\usage{
DynComm.communitiesEdgeCount(dyncomm)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the 
function call}
}
\value{
the number of community to community edges in the graph
}
\description{
This method returns the number of community to community edges in the graph
from the selected post processing algorithm or the main algorithm, after the 
last iteration.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$communitiesEdgeCount()
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
communitiesEdgeCount(dc)
}

}
\seealso{
\code{\link{DynComm}} , \code{\link{postProcess}}
}
\author{
poltergeist0
}
