% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{DynComm$community}
\alias{DynComm$community}
\alias{community}
\title{community(vertex)}
\usage{
community(vertex)
}
\arguments{
\item{vertex}{The name of the intended vertex}
}
\value{
an unsigned integer with the community of the given vertex
}
\description{
Get the community of the given vertex after the last iteration.
}
\examples{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$community(8)

}
