% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{community}
\alias{community}
\title{community(vertex)}
\usage{
community(vertex)
community(dyncomm,vertex)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the 
function call}

\item{vertex}{The name of the intended vertex}
}
\value{
an unsigned integer with the community of the given vertex
}
\description{
Get the community of the given vertex after the last iteration.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$community(8)
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
community(dc,8)
}

}
\author{
poltergeist0
}
