% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{communityCount}
\alias{communityCount}
\title{communityCount()}
\usage{
communityCount()
communityCount(dyncomm)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the 
function call}
}
\value{
an unsigned integer value with the number of communities
}
\description{
Get the number of communities after the last iteration of the algorithm.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$communityCount()
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
communityCount(dc)
}

}
\author{
poltergeist0
}
