% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{DynComm$communityEdgeWeight}
\alias{DynComm$communityEdgeWeight}
\alias{communityEdgeWeight}
\title{communityEdgeWeight(source,destination)}
\usage{
communityEdgeWeight(source,destination)
}
\arguments{
\item{source}{The name of the source vertex that is part of the edge}

\item{destination}{The name of the destination vertex that is part of the edge}
}
\value{
a floating point number with the weight
}
\description{
Get the weight of the edge that goes from source to destination after the last iteration.
}
\examples{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$communityEdgeWeight(12,42)

}
