% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{communityMapping}
\alias{communityMapping}
\title{communityMapping(differential)}
\usage{
communityMapping(differential)
communityMapping(dyncomm,differential)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the 
function call}

\item{differential}{If TRUE, only values that have changed in the latest run 
will be returned}

\item{file}{If given, outputs the community mapping to the given file instead 
of the console}
}
\value{
a matrix with either the community mapping or a boolean value
}
\description{
Get the community mapping for all communities after the last iteration.
}
\details{
If file is not given, returns a two column matrix with vertices in 
the first column and the communities in the second.

If file is given, returns a single row, single column matrix with TRUE or 
FALSE, depending whether if writing to file succeeded or failed, respectively.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$communityMapping()
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
communityMapping(dc,file="")
}

}
\author{
poltergeist0
}
