% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{communityTotalWeight}
\alias{communityTotalWeight}
\title{communityTotalWeight(community)}
\usage{
communityTotalWeight(community)
communityTotalWeight(dyncomm,community)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the 
function call}

\item{community}{The name of the intended community}
}
\value{
a floating point number with the weight
}
\description{
Get the sum of weights of all edges of the given community after the last 
iteration.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$communityTotalWeight(1)
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
communityTotalWeight(dc,1)
}

}
\author{
poltergeist0
}
