% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{communityVertexCount}
\alias{communityVertexCount}
\alias{communityNodeCount}
\title{communityVertexCount(community)}
\usage{
communityVertexCount(community)
communityVertexCount(dyncomm,community)
communityNodeCount(community)
communityNodeCount(dyncomm,community)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the function call}

\item{community}{The name of the intended community}
}
\value{
an unsigned integer with the number of vertices in the given community
}
\description{
Get the amount of vertices in the given community after the last iteration.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$communityVertexCount(3)
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
communityVertexCount(dc,3)
}

}
\author{
poltergeist0
}
