% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{communityVertexCount}
\alias{communityVertexCount}
\alias{communityVertexCount.DynComm}
\alias{communityNodeCount}
\alias{communityNodeCount.DynComm}
\title{communityVertexCount}
\usage{
communityVertexCount(x,community, ...)
communityNodeCount(x,community, ...)
}
\arguments{
\item{x}{A DynComm object, if not using the inline version of the
function call}

\item{community}{The name of the intended community}

\item{...}{Any future additional parameters}
}
\value{
an unsigned integer with the number of vertices in the given community
}
\description{
Get the amount of vertices in the given community from the selected post
processing algorithm or the main algorithm, after the last iteration.
}
\examples{
\donttest{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$communityVertexCount(3)
}
\donttest{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
communityVertexCount(dc,3)
}

}
\seealso{
\code{\link{DynComm}} , \code{\link{postProcess}}
}
\author{
poltergeist0
}
