% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{package}
\name{DynComm-package}
\alias{DynComm-package}
\title{DynComm: Dynamic Network Communities Detection}
\description{
This package is used for evolving network analysis regarding community detection.
Implements several algorithms that calculate communities for graphs whose nodes and edges change over time.
Edges, which can have new nodes, can be added or deleted.
Changes in the communities are calculated without recalculating communities for the entire graph.
}
