% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{class}
\name{DynComm}
\alias{DynComm}
\alias{Dyncomm}
\alias{dyncomm}
\title{DynComm}
\usage{
DynComm(Algorithm,Criterion,Parameters)
}
\arguments{
\item{Algorithm}{One of the available ALGORITHM See \code{\link{ALGORITHM}}}

\item{Criterion}{One of the available CRITERION. See \code{\link{CRITERION}}}

\item{Parameters}{A two column matrix defining additional parameters. See
the PARAMETERS section on this page}
}
\value{
\code{DynComm} object
}
\description{
Provides a single interface for all algorithms in the different 
languages.
}
\details{
Includes methods to get results of processing and to interact with the 
vertices, edges and communities.
Provided methods to return information on the graph are divided into two 
layers. A lower level layer that interacts with vertices and how they 
connect. And a higher level layer that interacts with communities and how 
they connect.
Besides the main algorithm, also accepts post processing algorithms.
}
\section{PARAMETERS}{

A two column matrix defining additional parameters to be passed to the
selected ALGORITHM and CRITERION.
The first column names the parameter and the second defines its value.
\describe{
  \item{
  -c
  }{
  Owsinski-Zadrozny quality function parameter. Values [0.0:1.0]. Default: 0.5
  }
  \item{
  -k
  }{
  Shi-Malik quality function kappa_min value. Value > 0 . Default 1
  }
  \item{
  -w
  }{
  Treat graph as weighted. In other words, do not ignore weights for edges 
  that define them when inserting edges in the graph.
  A weight of exactly zero removes the edge instead of inserting so its
  weight is never ignored.
  Without this parameter defined or for edges that do not have a weight defined, 
  edges are assigned the default value of 1 (one).
  As an example, reading from a file may define weights (a third column) for
  some edges (defined in rows, one per row) and not for others. With this
  parameter defined, the edges that have weights that are not exactly zero,
  have their weight replaced by the default value.
  }
  \item{
  -e
  }{
  Stops when, on a cycle of the algorithm, the quality is increased by less 
  than the value given in this parameter.
  }
}
}

\section{Methods}{

\describe{

  \item{postProcess(actions)}{Set a list of post processing steps. See \code{\link{postProcess}}}
  

  \item{select(postProcessing)}{Select between getting the results of the algorithm or one of the post processing steps. See \code{\link{select}}}
  

  \item{results(differential)}{Get additional results of the algorithm or the currently selected post processing steps. See \code{\link{results}}}
  

  \item{addRemoveEdges(graphAddRemove)}{Add and remove edges read from a matrix or file. See \code{\link{addRemoveEdges}}}
  

  \item{addRemove(graphAddRemove)}{Alias for addRemoveEdges(). See \code{\link{addRemoveEdges}}}
  

  \item{add(graphAddRemove)}{Alias for addRemoveEdges(). See \code{\link{addRemoveEdges}}}
  

  \item{quality()}{Get the quality measurement of the graph after the last iteration. See \code{\link{quality}}}
  

  \item{communityCount()}{Get the number of communities after the last iteration. See \code{\link{communityCount}}}
  

  \item{communities()}{Get all communities after the last iteration. See \code{\link{communities}}}
  

  \item{communityNeighbours(community)}{Get the neighbours of the given community after the last iteration. See \code{\link{communityNeighbours}}}
  

  \item{communityInnerEdgesWeight(community)}{Get the sum of weights of the inner edges of the given community after the last iteration. See \code{\link{communityInnerEdgesWeight}}}
  

  \item{communityTotalWeight(community)}{Get the sum of weights of all edges of the given community after the last iteration. See \code{\link{communityTotalWeight}}}
  

  \item{communityEdgeWeight(source,destination)}{Get the weight of the edge that goes from source community to destination community after the last iteration. See \code{\link{communityEdgeWeight}}}
  

  \item{communityVertexCount(community)}{Get the amount of vertices in the given community after the last iteration. See \code{\link{communityVertexCount}}}
  

  \item{communityNodeCount(community)}{Alias for communityVertexCount(). See \code{\link{communityVertexCount}}}
  

  \item{community(vertex)}{Get the community of the given vertex after the last iteration. See \code{\link{community}}}
  

  \item{vertexCount()}{Get the total number of vertices after the last iteration. See \code{\link{vertexCount}}}
  

  \item{nodesCount()}{Alias for vertexCount(). See \code{\link{vertexCount}}}
  

  \item{verticesAll()}{Get all vertices in the graph after the last iteration. See \code{\link{verticesAll}}}
  

  \item{nodesAll()}{Alias for verticesAll(). See \code{\link{verticesAll}}}
  

  \item{neighbours(vertex)}{Get the neighbours of the given vertex after the last iteration. See \code{\link{neighbours}}}
  

  \item{edgeWeight(source,destination)}{Get the weight of the edge that goes from source vertex to destination vertex after the last iteration. See \code{\link{edgeWeight}}}
  

  \item{edge(source,destination)}{Alias for edgeWeight(). See \code{\link{edgeWeight}}}
  

  \item{vertices(community)}{Get all vertices belonging to the given community after the last iteration. See \code{\link{vertices}}}
  

  \item{nodes(community)}{Alias for vertices(community). See \code{\link{vertices}}}
  

  \item{communityMapping(differential)}{Get the community mapping for all communities after the last iteration.See \code{\link{communityMapping}}}
  

  \item{time()}{Get the cumulative time spent on processing after the last iteration. See \code{\link{time}}}
  

}
}

\examples{
\dontrun{
Parameters<-matrix(c("-e","0.1"),1,2,TRUE)
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,Parameters)
dc$addRemoveEdgesFile("initial_graph.txt")
dc$addRemoveEdges(matrix(c(1,2,1,3,2,3,3,6,4,6,4,5,5,7,6,7),ncol=2,byrow=TRUE))
dc$communityCount()
dc$communities()
dc$communityNodeCount(1)
dc$vertices(1)
dc$communityMapping(TRUE)
dc$time()
dc$addRemoveEdgesFile("s0000000000.txt")
}

}
\author{
poltergeist0
}
