% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{class}
\name{DynComm}
\alias{DynComm}
\title{DynComm}
\usage{
\method{DynComm}{getAlgorithm}(Algorithm, Quality, parameters)
}
\arguments{
\item{x}{A description of the parameter 'x'. The
description can span multiple lines.}

\item{y}{A description of the parameter 'y'.}
}
\description{
This class provides a single interface for all algorithms in the different languages.
It provides methods to get results of processing and to interact with the nodes, edges and communities.
}
\section{A Custom Section}{


Text accompanying the custom section.
}

\examples{

## don't run this in calls to 'example(add_numbers)'
\dontrun{
   add_numbers(2, 3)
}

## don't test this during 'R CMD check'
\donttest{
   add_numbers(4, 5)
}


getAlgorithm

This method returns the algorithm with which this object was initialized.

dc<-DynComm(ALGORITHM$LOUVAIN,QUALITY$MODULARITY,parameters)
dc$getAlgorithm()

## don't run this in calls to 'example(add_numbers)'
\dontrun{
   add_numbers(2, 3)
}

## don't test this during 'R CMD check'
\donttest{
   add_numbers(4, 5)
}
}
