% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{edgeWeight}
\alias{edgeWeight}
\title{edgeWeight(source,destination)}
\usage{
edgeWeight(source,destination)
edgeWeight(dyncomm,source,destination)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the function call}

\item{source}{The name of the source vertex that is part of the edge}

\item{destination}{The name of the destination vertex that is part of the edge}
}
\value{
a floating point number with the weight
}
\description{
Get the weight of the edge that goes from source vertex to destination vertex after the last iteration.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$edgeWeight(12,42)
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
edgeWeight(dc,12,42)
}

}
\author{
poltergeist0
}
