% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{neighbours}
\alias{neighbours}
\alias{neighbours.DynComm}
\title{neighbours}
\usage{
neighbours.DynComm(dyncomm,vertex)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the 
function call}

\item{vertex}{The vertex to get neighbours from}
}
\value{
a matrix of all vertices in the graph that are neighbours of the 
given vertex and their edge weight
}
\description{
Get all neighbours (vertices connected through direct edges) of the given 
vertex in the graph from the selected post processing algorithm or the main 
algorithm, after the last iteration.
}
\examples{
\donttest{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$neighbours(vertex)
}
\donttest{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
neighbours(dc,vertex)
}

}
\seealso{
\code{\link{DynComm}} , \code{\link{postProcess}}
}
\author{
poltergeist0
}
