% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{DynComm$nodesAll}
\alias{DynComm$nodesAll}
\alias{nodesAll}
\title{nodesAll()}
\usage{
nodesAll()
}
\value{
a list of all nodes in the graph
}
\description{
Get all nodes in the graph after the last iteration.
}
\examples{
dc<-DynComm(ALGORITHM$LOUVAIN,QUALITY$MODULARITY,parameters)
dc$nodesAll()

}
