% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{quality}
\alias{quality}
\title{quality()}
\usage{
quality()
quality(dyncomm)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the function call}
}
\value{
a floating point number
}
\description{
Get the quality measurement of the graph after the last iteration of the 
algorithm.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$quality()
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
quality(dc)
}

}
\author{
poltergeist0
}
