% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{results}
\alias{results}
\title{results(differential)}
\usage{
results(differential)
results(dyncomm,differential)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the 
function call}

\item{differential}{If TRUE, only values that have changed in the latest run 
will be returned}
}
\value{
a two column matrix where, the first column is the name of the 
result and, the second column is its value.
}
\description{
This method returns additional results from the selected post processing
algorithm or the main algorithm. See \code{\link{select}} to know how to 
select an algorithm.
}
\details{
Additional results are any results other than those returned by other 
existing functions like \code{\link{quality}}, \code{\link{time}} and 
\code{\link{communityMapping}}.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$results()
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
results(dc)
}

}
\author{
poltergeist0
}
