% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{time}
\alias{time}
\title{time(differential=FALSE)}
\usage{
time()
time(dyncomm)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the 
function call}

\item{differential}{Select between differential and accumulated time.}
}
\value{
an unsigned integer with the total processing time
}
\description{
Get the time, in nanoseconds, spent on processing after the last iteration.
}
\details{
If the differential parameter is set, the time taken by the last iteration
will be returned. Otherwise, the default behaviour is to, return the 
accumulated time spent on processing since the creation of the DynComm 
object.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$time()
## 2.3
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
time(dc)
## 2.3
}

}
\author{
poltergeist0
}
