% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{vertexCount}
\alias{vertexCount}
\alias{vertexCount.DynComm}
\alias{nodeCount}
\alias{nodeCount.DynComm}
\title{vertexCount}
\usage{
vertexCount(x, ...)
nodeCount(x, ...)
}
\arguments{
\item{x}{A DynComm object, if not using the inline version of the
function call}

\item{...}{Any future additional parameters}
}
\value{
an unsigned integer with the number of vertices in the graph
}
\description{
Get the total number of vertices from the selected post processing
algorithm or the main algorithm, after the last iteration.
}
\details{
It can be useful since vertices can be added, if an edge being added has vertices
that do not exist in the graph, or removed, if they are not part of any edge after
removing an edge.
}
\examples{
\donttest{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$vertexCount()
}
\donttest{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
vertexCount(dc)
}

}
\seealso{
\code{\link{DynComm}} , \code{\link{postProcess}}
}
\author{
poltergeist0
}
