% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{vertices}
\alias{vertices}
\alias{nodes}
\title{vertices(community)}
\usage{
vertices(community)
vertices(dyncomm,community)
nodes(community)
nodes(dyncomm,community)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the 
function call}

\item{community}{The name of the intended community}
}
\value{
a list of vertices belonging to the given community
}
\description{
Get all vertices belonging to the given community after the last iteration.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$vertices(6)
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
vertices(dc,6)
}

}
\author{
poltergeist0
}
