% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynComm.R
\docType{methods}
\name{vertices}
\alias{vertices}
\alias{DynComm.vertices}
\alias{DynComm.nodes}
\title{vertices(community)}
\usage{
DynComm.vertices(dyncomm,community)
DynComm.nodes(dyncomm,community)
}
\arguments{
\item{dyncomm}{A DynComm object, if not using the inline version of the 
function call}

\item{community}{The name of the intended community}
}
\value{
a list of vertices belonging to the given community
}
\description{
Get all vertices belonging to the given community from the selected post 
processing algorithm or the main algorithm, after the last iteration.
}
\examples{
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
dc$vertices(6)
}
\dontrun{
dc<-DynComm(ALGORITHM$LOUVAIN,CRITERION$MODULARITY,parameters)
vertices(dc,6)
}

}
\seealso{
\code{\link{DynComm}} , \code{\link{postProcess}}
}
\author{
poltergeist0
}
